requirements.txt
# Post Service Python Dependencies
# MSA 환경에서 Post 서비스 구축을 위한 핵심 패키지들

# Flask 웹 프레임워크 (API 서버)
Flask==2.3.3

# 데이터베이스 연동 (MySQL)
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.1.0
PyMySQL==1.1.0
cryptography==41.0.7
boto3==1.34.0

# JWT 검증 (Cognito 토큰)
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# 이미지 처리 (S3 연동 시 필요)
Pillow==10.0.1

# HTTP 클라이언트 (MSA 서비스 간 통신)
requests

# AWS 연동 (S3, RDS 등)
boto3

# 개발 도구
python-dotenv==1.0.0

# API 관련 추가 패키지
Flask-CORS==4.0.0
flask-swagger-ui==4.11.1

# 테스트 관련 패키지
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# 코드 품질 관련 패키지
flake8==6.1.0
black==23.11.0
isort==5.12.0

